test/requirements.txt
pandas>=0.25
pytest-benchmark>=3.2
